+++ Editor/Content/ImportSettingsConfig/ImportingItem.cs
@@ -106,7 +106,7 @@ namespace Editor.Content.ImportSettingsConfig
         {
             ProgressMax = maxValue;
             ProgressValue = progress;
-            NormalizedValue = Math.Clamp(progress / maxValue, 0, 1);
+            NormalizedValue = maxValue > 0 ?  Math.Clamp(progress / maxValue, 0, 1) : 0.0;
         }
 
         private void UpdateTimer(object sender, EventArgs e)